Lighting and Skybox

+++ headers/input.h
@@ -4,7 +4,7 @@
 #include <GLFW/glfw3.h>
 #include <glm/glm.hpp>
 
-const float gravity = 9.80665;
+const double gravity = 9.80665;
 
 class Input {
 public:
